parse genomes input to collect_reference_bins as snake_dict
